JITM: use the A8c-Mc_Stats package
Use the A8c-Mc-Stats package to record stats instead of calling
Jetpack methods.

We're currently using the JETPACK__VERSION in the stat names. We'll
need to use something else when the package is used without Jetpack.
I've added a filter that plugins can use to change the version. The
default version is 'JITM-{package version}'. Of course, if multiple
plugins attempt to use the filter to set the version, they can
overwrite each other.